Update Hugo binary to 0.32.4
This is required for the OpenShift for Ops workshop to get source code
coloring to work correctly.

I did a spot check of other workshops and there was not anything obvious
that was "off" or looked out of place. The site generation also does not
give any errors so this should be a safe upgrade.