Fix Travis CI build
The Travis CI build broke on January 24 and started pushing the source
code for the redhatgov.io site to the master branch instead of the
generated site content. This caused the site to go down.

According to the docs, the GitHub Pages deployment feature in Travis CI
would use the content of the current directory as it's the content that
it would push to GitHub Pages. This has been working this way since the
start of the redhatgov.io site, but stopped working on January 24.

This commit changes back to the root directory of the code that was
cloned and then explicitly specifies the directory that contains the
generated content. After some testing, this appears to fix the problem.
It may be that Travis CI will fix it's bug at some point and we can go
back to how it was done prior, but this workaround solves the immediate
problem.