Fix events generation due to change in google-api-python-client
Previous builds were experiencing the following error:

Traceback (most recent call last):
  File "./import_events.py", line 13, in <module>
    from apiclient import discovery
  File "/usr/local/lib/python2.7/dist-packages/apiclient/__init__.py", line 11, in <module>
    'Previous version of google-api-python-client detected; due to a '
RuntimeError: Previous version of google-api-python-client detected; due to a packaging issue, we cannot perform an in-place upgrade. To repair, remove and reinstall this package, along with oauth2client and uritemplate. One can do this with pip via
  pip install -I google-api-python-client

This was due to a change in the google-api-python-client package. We got
bit by this because we did not specify exact versions of each package to
pull and therefore were always getting the latest.

I have added the install of oauth2client as it's now required by
google-api-python-client and also updated the requirements.txt file to
include exact versions to minimize the risk in the future.